Backend/Python/requirements.txt
xbox-webapi-python
appdirs>=1.4.4
ecdsa>=0.18.0
httpx>=0.24.0
ms_cv>=0.1.0
pydantic>=2.0.0
requests>=2.28.0